Extract KeepBadISBNDialog to its own file
